refactor(wizard): move config out
