+++ src/app/space/analyze/analyze-overview/analyze-overview.component.html
@@ -35,6 +35,6 @@
 <!-- Forge wizard modal -->
 <ng-template #addSpace>
     <flow-selector (onSelect)="selectFlow($event)" (onCancel)="closeModal()" [space]="space.attributes.name" *ngIf="!selectedFlow"></flow-selector>
-    <forge-wizard (onCancel)="closeModal()" *ngIf="selectedFlow === 'import'"></forge-wizard>
-    <forge-quickstart *ngIf="selectedFlow === 'quickstart'"></forge-quickstart>
+    <import-wizard (onCancel)="closeModal()" *ngIf="selectedFlow === 'import'"></import-wizard>
+    <quickstart-wizard (onCancel)="closeModal()" *ngIf="selectedFlow === 'quickstart'"></quickstart-wizard>
 </ng-template>
