+++ src/app/space/create/pipelines/pipelines.component.html
@@ -42,6 +42,7 @@
 
 <!-- Forge wizard modal -->
 <ng-template #updateSpace>
-  <flow-selector (onSelect)="selectFlow($event)" (onCancel)="closeModal()" [space]="space.attributes.name" *ngIf="!isFlowSelected"></flow-selector>
-  <forge-wizard (onCancel)="closeModal()" *ngIf="isFlowSelected"></forge-wizard>
+  <flow-selector (onSelect)="selectFlow($event)" (onCancel)="closeModal()" [space]="space.attributes.name" *ngIf="!selectedFlow"></flow-selector>
+  <import-wizard (onCancel)="closeModal()" *ngIf="selectedFlow === 'import'"></import-wizard>
+  <quickstart-wizard (onCancel)="closeModal()" *ngIf="selectedFlow === 'quickstart'"></quickstart-wizard>
 </ng-template>
